project/requirements.txt
yfinance 
datetime
plotly.graph_objs
requests
json
gudhi
pandas